requirements.txt
dj-database-url
Django
gunicorn
psycopg2-binary
wheel
whitenoise

python-decouple
django-debug-toolbar